+++ R/check_esiste.R
@@ -1,0 +1,3 @@
+check_esiste <- function(soluzione, risposta) {
+  !is.null(soluzione[risposta][[1]])
+}