+++ R/score.R
@@ -5,5 +5,5 @@
 #' @return 1 if TRUE, -1 if FALSE
 score <- function(x) {
   assertive::assert_is_a_bool(x)
-  (x - 0.5) * 2
+  as.integer((x - 0.5) * 2)
 }