+++ R/tdp_2018_2_a_checks.R
@@ -1,11 +1,15 @@
-#' tdp check 1
+#' tdp 2018, appello 2, tema a, check 1
 #'
-#' @param file_to_check [chr] name of a file which should be into the WD
+#' @param wd [chr] working directory
+#' @param file_to_check [chr] path of a file to check into the wd
 #'
 #' @return logical (invisibly)
 #' @export
-tdp1718_check_1 <- function(file_to_check = "death_ita.csv"
-) {
+tdp_2018_2_a_check_1 <- function(wd = getwd(), file_to_check = "death_ita.csv"){
+
+  old_wd <- setwd(wd)
+  on.exit(setwd(old_wd), add = TRUE)
+
   message(paste0(
     'Directory di lavoro impostata: ', getwd(), '.\n'
   ))
@@ -29,17 +33,22 @@ tdp1718_check_1 <- function(file_to_check = "death_ita.csv"
 
 
 
-
-
-
-#' tdp check 2
+#' tdp 2018, appello 2, tema a, check 2
+#'
+#' @param provided_names names included in the workspace
+#' @param required_names variable'names required
 #'
 #' @return logical (invisibly)
 #' @export
-tdp1718_check_2 <- function() {
-  required_names <- "death_ita"
-  names_ok <- exists(required_names, envir = parent.frame(), inherits = FALSE)
-
+tdp_2018_2_a_check_2 <- function(
+  provided_names = NULL,
+  required_names = "death_ita"
+) {
+  if (is.null(provided_names)) {
+    names_ok <- exists(required_names, envir = parent.frame(), inherits = FALSE)
+  } else {
+    names_ok <- required_names %in% provided_names
+  }
   if (!all(names_ok)) {
     message(paste0(
       "L'oggetto richiesto '", required_names[!names_ok], "' manca."
@@ -71,12 +80,15 @@ tdp1718_check_2 <- function() {
 
 
 
-
-#' tdp check 3
+#' tdp 2018, appello 2, tema a, check 3
+#'
+#' @param data a list of data to check
+#' @param reference a list of real data
 #'
 #' @return logical (invisibly)
 #' @export
-tdp1718_check_3 <- function() {
+tdp_2018_2_a_check_3 <- function(data, reference) {
+
   required_names <- c("numero_righe", "numero_colonne", "nomi_colonne")
   names_ok <- purrr::map_lgl(required_names, exists)
 
@@ -88,9 +100,9 @@ tdp1718_check_3 <- function() {
   }
 
   message(paste0(
-    'Il numero di righe della base di dati è: ',   numero_righe,   ';\n',
-    'Il numero di colonne della base di dati è: ', numero_colonne, ';\n',
-    'I nomi delle colonne della base di dati sono: ',
+    'Numero di righe proposte: ',   numero_righe,   ';\n',
+    'Numero di colonne proposte: ', numero_colonne, ';\n',
+    'Nomi proposti per le colonne: ',
     paste(nomi_colonne, collapse = ", "), '.\n'
   ))
 
@@ -105,7 +117,6 @@ tdp1718_check_3 <- function() {
     message('Almeno una risposta sembra non essere corretta...\n')
     return(invisible(FALSE))
   }
-
 }
 
 
@@ -116,14 +127,13 @@ tdp1718_check_3 <- function() {
 
 
 
-
-#' tdp check 4
+#' tdp 2018, appello 2, tema a, check 4
 #'
 #' @param main_df a data.frame
 #'
 #' @return logical (invisibly)
 #' @export
-tdp1718_check_4 <- function(main_df = death_ita) {
+tdp_2018_2_a_check_4 <- function(main_df) {
   if ('cause_of_death' %in% colnames(main_df)) {
     message(
       'la colonna `cause_of_death` sembra essere ancora presente nel dataset.\n'
@@ -153,7 +163,11 @@ tdp1718_check_4 <- function(main_df = death_ita) {
 
 
 
-#' tdp check 5
+
+
+
+
+#' tdp 2018, appello 2, tema a, check 5
 #'
 #' @param text_var [chr] name of the variable to transform
 #' @param .fun the function requested for the transformation. Note, the
@@ -162,10 +176,16 @@ tdp1718_check_4 <- function(main_df = death_ita) {
 #'
 #' @return logical (invisibly)
 #' @export
-tdp1718_check_5 <- function(text_var = "causa_del_decesso", .fun = toupper) {
+tdp_2018_2_a_check_5 <- function(main_df, .fun = tolower
+) {
+  if (is.null(main_df[["causa_del_decesso"]])) {
+    message("L'esercizio non sembra essere corretto...\n")
+    message("variabile causa_del_decesso non presente nel dataset")
+    return(invisible(FALSE))
+  }
   if (all(
-    death_ita[[text_var]] ==
-    .fun(death_ita[[text_var]])
+    main_df[["causa_del_decesso"]] ==
+    .fun(read.csv("death_ita.csv", stringsAsFactors = FALSE)[["cause_of_death"]])
   )) {
     message("L'esercizio sembra essere corretto.\n")
     return(invisible(TRUE))
@@ -184,16 +204,27 @@ tdp1718_check_5 <- function(text_var = "causa_del_decesso", .fun = toupper) {
 
 
 
-#' tdp check 6
+
+
+
+#' tdp 2018, appello 2, tema a, check 6
 #'
-#' @param text_var [chr] name of the variable to transform
+#' @param main_df [data.frame] main dataframe
+#' @param new_value [chr] character used to substitute the white spaces
 #'
 #' @return logical (invisibly)
 #' @export
-tdp1718_check_6 <- function(text_var = "causa_del_decesso") {
+tdp_2018_2_a_check_6 <- function(main_df, new_value = '') {
+  if (is.null(main_df[["causa_del_decesso"]])) {
+    message("L'esercizio non sembra essere corretto...\n")
+    message("variabile causa_del_decesso non presente nel dataset")
+    return(invisible(FALSE))
+  }
   if (all(
-    death_ita[[text_var]] ==
-    gsub(" ", "_", death_ita[[text_var]])
+    tolower(main_df[["causa_del_decesso"]]) ==
+    tolower(gsub(" ", new_value,
+      read.csv("death_ita.csv", stringsAsFactors = FALSE)[["cause_of_death"]]
+    ))
   )) {
     message("L'esercizio sembra essere corretto.\n")
     return(invisible(TRUE))
@@ -212,28 +243,30 @@ tdp1718_check_6 <- function(text_var = "causa_del_decesso") {
 
 
 
-#' tdp check 7
+
+
+
+
+#' tdp 2018, appello 2, tema a, check 7
+#'
+#' @param main_df [data.frame] main dataframe
 #'
 #' @return logical (invisibly)
 #' @export
-tdp1718_check_7 <- function() {
-  if (!exists('max_perc_death')) {
-    message("Il dataframe `max_perc_death` sembra non essere stato creato...\n")
-    return(invisible(FALSE))
-  } else if (!is.data.frame(max_perc_death)) {
-    message("`max_perc_death` sembra non essere un dataframe...\n")
-    return(invisible(FALSE))
-  } else if (
-    nrow(max_perc_death) != 1 ||
-    max_perc_death[['causa_del_decesso']] != 'PREMATURITY' ||
-    max_perc_death[['anno']] != 2000 ||
-    (!max_perc_death[['mesi']] %in% c('0-1', 'neonato'))
-  ) {
-    message('Il risultato sembra non essere quello corretto...\n')
-    return(invisible(FALSE))
-  } else {
+tdp_2018_2_a_check_7 <- function(main_df) {
+  if (
+    all.equal(
+      main_df[["percentuale"]],
+      read.csv("death_ita.csv", stringsAsFactors = FALSE) %>%
+        `[[`("percentuale") %>%
+        `[`(. > 0)
+    )
+  ){
     message('Il risultato sembra corretto.\n')
     return(invisible(TRUE))
+  } else {
+    message('Il risultato sembra non essere quello corretto...\n')
+    return(invisible(FALSE))
   }
 }
 
@@ -244,50 +277,39 @@ tdp1718_check_7 <- function() {
 
 
 
-
-
-#' tdp check 8
+#' tdp 2018, appello 2, tema a, check 8
 #'
-#' @param df_name [chr] the name of the principal data.frame
+#' @param main_df [chr] main data.frame
 #'
 #' @return logical (invisibly)
 #' @export
-tdp1718_check_8 <- function(df_name   = "death_ita",
+tdp_2018_2_a_check_8 <- function(main_df,
                             ord_one   = 'percentuale',
                             ord_two   = 'anno',
                             ord_three = 'mesi'
 ) {
-  reference <- read.csv(
-    system.file("exams/tdp1718/death_ita.csv", package = "rexams")
-  )[-1]
+  reference <- read.csv("death_ita.csv", stringsAsFactors = FALSE)[-1]
   reference[["mesi"]] <- factor(reference[["mesi"]],
     levels = c("0-1", "2-59", "0-59"),
     labels = c("neonato", "fanciullo", "prescolare")
   )
   reference <- reference[
     order(reference$percentuale, reference$anno, as.character(reference$mesi)),
-  ]
+    ]
 
-  if (!exists(df_name)) {
-    message("The main dataframe seams not imported...\n")
-    return(invisible(FALSE))
-  }
-
-  if (!is.data.frame(death_ita)) {
-    message("The main dataframe seams not to be a dataframe...\n")
+  if (!is.data.frame(main_df)) {
+    message("The object provided seams not to be a dataframe...\n")
     return(invisible(FALSE))
   }
 
   if (any(
-    purrr::map_lgl(c(ord_one, ord_two, ord_three),
-      ~is.null(death_ita[[.]])
-    )
+    purrr::map_lgl(c(ord_one, ord_two, ord_three), ~is.null(main_df[[.]]))
   )) {
     message("The main dataframe's variables seams not to be correct...\n")
     return(invisible(FALSE))
   }
 
-  to_check <- death_ita[-1] %>% as.data.frame()
+  to_check <- main_df[-1] %>% as.data.frame()
   to_check <- to_check[
     order(to_check[[ord_one]], to_check[[ord_two]], as.character(to_check[[ord_three]])),
     ]
@@ -297,7 +319,7 @@ tdp1718_check_8 <- function(df_name   = "death_ita",
   if (
     isTRUE(all.equal(reference, to_check)) &&
     setequal(
-      levels(death_ita[[ord_three]]),
+      levels(main_df[[ord_three]]),
       c("neonato", "prescolare", "fanciullo")
     )
   ) {
@@ -315,37 +337,37 @@ tdp1718_check_8 <- function(df_name   = "death_ita",
 
 
 
-
-
-
-#' tdp check 9
+#' tdp 2018, appello 2, tema a, check 9
 #'
-#' @param matrix_name [chr] the name of the requested matrix
+#' @param main_table [matrix] the requested matrix
 #'
 #' @return logical (invisibly)
 #' @export
-tdp1718_check_9 <- function(matrix_name = "median_p_death_causes") {
-  if (!exists(matrix_name)) {
-    message(
-      "La tabella `median_p_death_causes` sembra non essere stata creata...\n"
-    )
-    return(invisible(FALSE))
-  } else if (!is.matrix(median_p_death_causes)) {
-    message("`median_p_death_causes` sembra non essere una tabella/matrice.\n")
+tdp_2018_2_a_check_9 <- function(main_table) {
+  if (!is.matrix(main_table)) {
+    message("L'oggetto fornito sembra non essere una tabella/matrice.\n")
     return(invisible(FALSE))
   }
 
+  ref <- read.csv("death_ita.csv", stringsAsFactors = FALSE)
+  ref[["mesi"]] <- factor(ref[["mesi"]],
+    levels = c("0-1", "2-59", "0-59"),
+    labels = c("neonato", "fanciullo", "prescolare")
+  )
+
   reference <- tapply(
-      X     = death_ita$percentuale,
-      INDEX = list(death_ita$causa_del_decesso, death_ita$mesi),
-      FUN   = median, na.rm = TRUE
-    ) %>%
+    X     = ref$percentuale,
+    INDEX = list(ref$cause_of_death, ref$mesi),
+    FUN   = max, na.rm = TRUE
+  ) %>%
     as.data.frame() %>%
+    dplyr::select(neonato, prescolare, fanciullo) %>%
     dplyr::arrange(neonato, prescolare, fanciullo) %>%
     as.matrix() %>%
     as.vector()
 
-  to_check <- as.data.frame(median_p_death_causes) %>%
+  to_check <- as.data.frame(main_table) %>%
+    dplyr::select(neonato, prescolare, fanciullo) %>%
     dplyr::arrange(neonato, prescolare, fanciullo) %>%
     as.matrix() %>%
     as.vector()
@@ -368,13 +390,22 @@ tdp1718_check_9 <- function(matrix_name = "median_p_death_causes") {
 
 
 
-#' tdp check 10
+
+
+
+
+
+
+
+#' tdp 2018, appello 2, tema a, check 10
+#'
+#' @param main_data A list with one or two dataframe
 #'
 #' @return logical (invisibly)
 #' @export
-tdp1718_check_10 <- function() {
-  df_requested <- c('bot_fan_2000', 'bot_fan_2016')
-  df_ok <- purrr::map_lgl(df_requested, exists)
+tdp_2018_2_a_check_10 <- function(main_data) {
+  df_requested <- c('top_fan_2000', 'top_fan_2016')
+  df_ok <- purrr::map_lgl(df_requested, ~ . %in% names(main_data))
   df_provided <- df_requested[df_ok]
 
 
@@ -386,24 +417,27 @@ tdp1718_check_10 <- function() {
   first <- TRUE
   score <- 0L
 
+  top_fan_2000 <- main_data[["top_fan_2000"]]
+  top_fan_2016 <- main_data[["top_fan_2016"]]
+
   try(silent = TRUE, {
-    if (!is.data.frame(bot_fan_2000)) {
-      message('`bot_fan_2000` sembra non essere un dataframe...\n')
+    if (!is.data.frame(top_fan_2000)) {
+      message('`top_fan_2000` sembra non essere un dataframe...\n')
       mark <- FALSE
     } else if (
-        bot_fan_2000[['causa_del_decesso']][1:2] ==
-          c('PERTUSSIS', 'TETANUS')          &&
+      top_fan_2000[['causa_del_decesso']][1:2] ==
+      c('prematurity', 'congenitalanomalies')          &&
 
-        bot_fan_2000[['percentuale']][2:3] == c(0.0, 0.0) &&
+      top_fan_2000[['percentuale']][2:3] == c(28.7, 13.5) &&
 
-        all(bot_fan_2000[['anno']] == 2000)                 &&
+      all(top_fan_2000[['anno']] == 2000)                 &&
 
-        all(bot_fan_2000[['mesi']] %in% c('fanciullo', '1-59'))
+      all(top_fan_2000[['mesi']] %in% c('neonato', '0-1'))
     ) {
-      message('`bot_fan_2000` sembra essere corretto.\n')
+      message('`top_fan_2000` sembra essere corretto.\n')
       mark <- TRUE
     } else {
-      message('`bot_fan_2000` sembra non essere corretto...\n')
+      message('`top_fan_2000` sembra non essere corretto...\n')
       mark <- FALSE
     }
 
@@ -416,23 +450,23 @@ tdp1718_check_10 <- function() {
 
   mark <- FALSE
   try(silent = TRUE, {
-    if (!is.data.frame(bot_fan_2016)) {
-      message('`bot_fan_2016` sembra non essere un dataframe...\n')
+    if (!is.data.frame(top_fan_2016)) {
+      message('`top_fan_2016` sembra non essere un dataframe...\n')
       mark <- FALSE
     } else if (
-      bot_fan_2016[['causa_del_decesso']][1:2] ==
-      c('HIV/AIDS', 'TETANUS')            &&
+      top_fan_2016[['causa_del_decesso']][1:2] ==
+      c('prematurity', 'congenitalanomalies')            &&
 
-      bot_fan_2016[['percentuale']][2:3] == c(0.0, 0.0) &&
+      top_fan_2016[['percentuale']][2:3] == c(23.4, 21.7) &&
 
-      all(bot_fan_2016[['anno']] == 2016)                 &&
+      all(top_fan_2016[['anno']] == 2016)                 &&
 
-      all(bot_fan_2016[['mesi']]  %in% c('fanciullo', '1-59'))
+      all(top_fan_2016[['mesi']]  %in% c('neonato', '0-1'))
     ) {
-      message('`bot_fan_2016` sembra essere corretto.\n')
+      message('`top_fan_2016` sembra essere corretto.\n')
       mark <- TRUE
     } else {
-      message('`bot_fan_2016` sembra non essere corretto...\n')
+      message('`top_fan_2016` sembra non essere corretto...\n')
       mark <- FALSE
     }
 
@@ -443,76 +477,10 @@ tdp1718_check_10 <- function() {
     }
   })
 
-  if (score < 0) {
+  if (score < 0L) {
     score <- score - first
   }
 
   invisible(score)
 }
 
-
-
-
-
-
-
-
-
-
-#' tdp score overall
-#'
-#' @return the final score
-#' @export
-score_my_exam <- function() {
-  suppressMessages({
-    partial_score <-
-      tdp1718_check_3() +
-      tdp1718_check_4() +
-      tdp1718_check_5() +
-      tdp1718_check_6() +
-      2L * tdp1718_check_7() +
-      2L * tdp1718_check_8() +
-      2L * tdp1718_check_9() +
-      tdp1718_check_10()
-  })
-
-  if (partial_score == 0L) {
-    final_score <-
-      score(tdp1718_check_1) +
-      score(tdp1718_check_2)
-  } else {
-    suppressMessages({
-      final_score <-
-        rexams:::score(tdp1718_check_3()) +
-        rexams:::score(tdp1718_check_4()) +
-        rexams:::score(tdp1718_check_5()) +
-        rexams:::score(tdp1718_check_6()) +
-        2L * rexams:::score(tdp1718_check_7()) +
-        2L * rexams:::score(tdp1718_check_8()) +
-        2L * rexams:::score(tdp1718_check_9()) +
-        tdp1718_check_10() + 2L
-    })
-  }
-
-  message(paste0(
-    "Punteggio finale stimato per questo esame: ", final_score, ".\n"
-  ))
-}
-
-
-
-
-
-
-
-
-
-#' Submit exam
-#'
-#' @param x filename
-#'
-#' @export
-submit_my_exam <- function(x = 'tdp1718.R') {
-  rmarkdown::render(x, encoding = "UTF-8", envir = new.env())
-  utils::browseURL(stringr::str_replace_all(x, "(.+)\\..+?$", "\\1\\.nb\\.html"))
-}